Refactor: Move NetworkIdWithKeyPair into bisq.network.identity
