+++ bonded_roles/src/main/java/bisq/bonded_roles/registration/BondedRoleRegistrationService.java
@@ -26,8 +26,8 @@ import bisq.common.observable.collection.ObservableSet;
 import bisq.network.NetworkService;
 import bisq.network.common.AddressByTransportTypeMap;
 import bisq.network.identity.NetworkId;
+import bisq.network.identity.NetworkIdWithKeyPair;
 import bisq.network.identity.TorIdentity;
-import bisq.network.p2p.vo.NetworkIdWithKeyPair;
 import bisq.security.DigestUtil;
 import lombok.extern.slf4j.Slf4j;
