+++ chat/src/main/java/bisq/chat/priv/PrivateChatChannelService.java
@@ -23,9 +23,9 @@ import bisq.i18n.Res;
 import bisq.network.NetworkService;
 import bisq.network.SendMessageResult;
 import bisq.network.identity.NetworkId;
+import bisq.network.identity.NetworkIdWithKeyPair;
 import bisq.network.identity.TorIdentity;
 import bisq.network.p2p.services.confidential.MessageListener;
-import bisq.network.p2p.vo.NetworkIdWithKeyPair;
 import bisq.persistence.PersistableStore;
 import bisq.security.pow.ProofOfWorkService;
 import bisq.user.UserService;
