+++ user/src/main/java/bisq/user/identity/UserIdentity.java
@@ -19,7 +19,7 @@ package bisq.user.identity;
 
 import bisq.common.proto.Proto;
 import bisq.identity.Identity;
-import bisq.network.p2p.vo.NetworkIdWithKeyPair;
+import bisq.network.identity.NetworkIdWithKeyPair;
 import bisq.user.profile.UserProfile;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
